+++ b2sdk/file_lock.py
@@ -161,6 +161,17 @@ class FileRetentionSetting:
             "retainUntilTimestamp": self.retain_until,
         }
 
+    def as_dict_with_auth(self):
+        if self == UNKNOWN_FILE_RETENTION_SETTING:
+            return {
+                "isClientAuthorizedToRead": False,
+                "value": None,
+            }
+        return {
+            "isClientAuthorizedToRead": True,
+            "value": self.as_dict(),
+        }
+
     def add_to_to_upload_headers(self, headers):
         if self.mode is RetentionMode.UNKNOWN:
             raise ValueError('cannot use an unknown file retention setting in requests')
@@ -246,6 +257,17 @@ class LegalHold(enum.Enum):
         elif self.is_unknown():
             return self.__class__.UNKNOWN.value
         raise ValueError('Unrepresentable value')
+
+    def as_dict_with_auth(self):
+        if self == self.__class__.UNKNOWN:
+            return {
+                "isClientAuthorizedToRead": False,
+                "value": None,
+            }
+        return {
+            "isClientAuthorizedToRead": True,
+            "value": self.to_dict_repr(),
+        }
 
 
 class BucketRetentionSetting: