+++ b2sdk/file_version.py
@@ -86,16 +86,21 @@ class FileVersion:
         else:
             self.mod_time_millis = self.upload_timestamp
 
-    def as_dict(self):
+    def as_dict(self, include_auth=False):
         """ represents the object as a dict which looks almost exactly like the raw api output for upload/list """
         result = {
             'fileId': self.id_,
             'fileName': self.file_name,
             'fileInfo': self.file_info,
-            'legalHold': self.legal_hold.to_dict_repr() if self.legal_hold is not None else None,
             'serverSideEncryption': self.server_side_encryption.as_dict(),
-            'fileRetention': self.file_retention.as_dict(),
         }
+
+        if include_auth:
+            result['legalHold'] = self.legal_hold.as_dict_with_auth()
+            result['fileRetention'] = self.file_retention.as_dict_with_auth()
+        else:
+            result['legalHold'] = self.legal_hold.to_dict_repr()
+            result['fileRetention'] = self.file_retention.as_dict()
 
         if self.size is not None:
             result['size'] = self.size