+++ b2sdk/v1/bucket.py
@@ -8,7 +8,7 @@
 #
 ######################################################################
 
-from .file_version import FileVersionInfoFactory
+from .file_version import FileVersionInfo, FileVersionInfoFactory
 from typing import Optional
 from b2sdk import _v2 as v2
 from b2sdk.utils import validate_b2_file_name
@@ -16,6 +16,7 @@ from b2sdk.utils import validate_b2_file_name
 
 # Overridden to retain the obsolete copy_file and start_large_file methods
 # and to return old style FILE_VERSION_FACTORY attribute
+# and to to adjust to old style B2Api.get_file_info return type
 class Bucket(v2.Bucket):
     FILE_VERSION_FACTORY = staticmethod(FileVersionInfoFactory)
 
@@ -88,6 +89,14 @@ class Bucket(v2.Bucket):
             file_retention=file_retention,
             legal_hold=legal_hold,
         )
+
+    def get_file_info_by_id(self, file_id: str) -> FileVersionInfo:
+        """
+        Gets a file version's by ID.
+
+        :param str file_id: the id of the file who's info will be retrieved.
+        """
+        return self.api.file_version_factory.from_api_response(self.api.get_file_info(file_id))
 
 
 class BucketFactory(v2.BucketFactory):