+++ CHANGELOG.md
@@ -6,6 +6,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Changed
+* `B2Api.get_file_info` returns a `FileVersion` object in v2
+
 ## [1.9.0] - 2021-06-07
 
 ### Added